refactoring, restyling proposal page
