+++ apps/app/src/components/decisions/types.ts
@@ -1,0 +1,14 @@
+export interface ProcessPhase {
+  id: string;
+  name: string;
+  description?: string;
+  phase?: {
+    startDate?: string;
+    endDate?: string;
+    sortOrder?: number;
+  };
+  type?: 'initial' | 'intermediate' | 'final';
+  config?: {
+    allowProposals?: boolean;
+  };
+}
